pkg/destroy: align destroyers code base
move aws destroyer init function to subpackage
move New to the top for better readability
