+++ pkg/destroy/aws/aws.go
@@ -22,6 +22,9 @@ import (
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/util/wait"
 
+	awssession "github.com/openshift/installer/pkg/asset/installconfig/aws"
+	"github.com/openshift/installer/pkg/destroy"
+	"github.com/openshift/installer/pkg/types"
 	"github.com/openshift/installer/pkg/version"
 )
 
@@ -61,6 +64,27 @@ type ClusterUninstaller struct {
 	// new session will be created based on the usual credential
 	// configuration (AWS_PROFILE, AWS_ACCESS_KEY_ID, etc.).
 	Session *session.Session
+}
+
+// New returns an AWS destroyer from ClusterMetadata.
+func New(logger logrus.FieldLogger, metadata *types.ClusterMetadata) (destroy.Destroyer, error) {
+	filters := make([]Filter, 0, len(metadata.ClusterPlatformMetadata.AWS.Identifier))
+	for _, filter := range metadata.ClusterPlatformMetadata.AWS.Identifier {
+		filters = append(filters, filter)
+	}
+
+	session, err := awssession.GetSession()
+	if err != nil {
+		return nil, err
+	}
+
+	return &ClusterUninstaller{
+		Filters:   filters,
+		Region:    metadata.ClusterPlatformMetadata.AWS.Region,
+		Logger:    logger,
+		ClusterID: metadata.InfraID,
+		Session:   session,
+	}, nil
 }
 
 func (o *ClusterUninstaller) validate() error {
