+++ pkg/destroy/aws/doc.go
@@ -1,0 +1,2 @@
+// Package aws provides a cluster-destroyer for AWS clusters.
+package aws
