+++ pkg/destroy/aws/register.go
@@ -1,0 +1,7 @@
+package aws
+
+import "github.com/openshift/installer/pkg/destroy"
+
+func init() {
+	destroy.Registry["aws"] = New
+}
