+++ pkg/destroy/bootstrap/doc.go
@@ -1,0 +1,2 @@
+// Package bootstrap provides a cluster-destroyer for Bootstrap node
+package bootstrap
