+++ pkg/destroy/libvirt/libvirt.go
@@ -47,6 +47,15 @@ type ClusterUninstaller struct {
 	Logger     logrus.FieldLogger
 }
 
+// New returns libvirt Uninstaller from ClusterMetadata.
+func New(logger logrus.FieldLogger, metadata *types.ClusterMetadata) (destroy.Destroyer, error) {
+	return &ClusterUninstaller{
+		LibvirtURI: metadata.ClusterPlatformMetadata.Libvirt.URI,
+		Filter:     ClusterIDPrefixFilter(metadata.InfraID),
+		Logger:     logger,
+	}, nil
+}
+
 // Run is the entrypoint to start the uninstall process.
 func (o *ClusterUninstaller) Run() error {
 	conn, err := libvirt.NewConnect(o.LibvirtURI)
@@ -209,13 +218,4 @@ func deleteNetwork(conn *libvirt.Connect, filter filterFunc, logger logrus.Field
 		logger.WithField("network", nName).Info("Deleted network")
 	}
 	return nil
-}
-
-// New returns libvirt Uninstaller from ClusterMetadata.
-func New(logger logrus.FieldLogger, metadata *types.ClusterMetadata) (destroy.Destroyer, error) {
-	return &ClusterUninstaller{
-		LibvirtURI: metadata.ClusterPlatformMetadata.Libvirt.URI,
-		Filter:     ClusterIDPrefixFilter(metadata.InfraID),
-		Logger:     logger,
-	}, nil
 }
